requirements.txt
python-telegram-bot==20.8
spotipy
yt-dlp
youtube-search-python
google-api-python-client
python-dotenv
requests
beautifulsoup4